Refactors document view into its own file with rom PR Fixes
